Manually implement Deserialize to enable json()
Reqwest includes the `json()` method on a response which uses Serde's
`DeserializeOwned` trait on whatever type you want to deserialize to.
So we make up some structs and an enum to hold the data from the json.
The enum is problematic because the moon phases in the json have spaces
in them, and enum variants are not allowed to contain spaces.  Manually
implementing `Deserialize` for the enum seems to be the most-elegant way
around that.